PR feedback. Move IdentityService from CommonServices and change the name of the contribution ID
