+++ src/Identities/index.ts
@@ -1,2 +1,3 @@
 export * from "./Identities";
 export * from "./IdentityProvider";
+export * from "./IdentityService";
